Fix a bug on `$.chunk`s starting with `<`
Previously, if a `chunk` started with the `<` immediately followed by a
line break, the chunk would continue onto the next line, which is
invalid.  Although none of the languages parsed by Neovim can have a
statement starting with a `<` AFAIK – making this a non-issue, I think
it's judicious to fix this now to avoid problems later.